Modified: Split config.cpp and bubber_manager.cpp as files were getting rather long.
